[kuduraft] [compression] abilty to compress messages read from the log
Summary:
Adds the ability to compress messages that are read from the log.
(1) Requests meant for proxy host is not compressed. This is because the
pyload is ignored for a proxy request and compression is wasteful
(2) A new flag in log-cache controls the feature so that it can be
enabled/disabled. This makes rollout and testing easier.
(3) Compression is done outside of LogCache::lock_ to avoid contending
with writes.

Test Plan:
Integration tests added in plugin and mtr in mysql

Reviewers: arahut

Subscribers:

Tasks:

Tags: